Customer API

+++ src/main/java/org/k9m/rental/persistence/repository/CustomerRepository.java
@@ -3,4 +3,4 @@ package org.k9m.rental.persistence.repository;
 import org.k9m.rental.persistence.model.CustomerDTO;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<CustomerDTO, Integer> {}
+public interface CustomerRepository extends JpaRepository<CustomerDTO, Long> {}
