+++ src/main/java/org/k9m/rental/persistence/repository/VehicleRepository.java
@@ -3,4 +3,4 @@ package org.k9m.rental.persistence.repository;
 import org.k9m.rental.persistence.model.CustomerDTO;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleRepository extends JpaRepository<CustomerDTO, Integer> {}
+public interface VehicleRepository extends JpaRepository<CustomerDTO, Long> {}
